feat: [ANDROAPP-5125] move RelationshipViewModel to tracker
