+++ app/src/main/java/org/dhis2/usescases/searchTrackEntity/SearchRepositoryImpl.java
@@ -12,9 +12,6 @@ import org.dhis2.commons.Constants;
 import org.dhis2.commons.data.EntryMode;
 import org.dhis2.commons.data.EventViewModel;
 import org.dhis2.commons.data.EventViewModelType;
-import org.dhis2.commons.data.RelationshipDirection;
-import org.dhis2.commons.data.RelationshipOwnerType;
-import org.dhis2.commons.data.RelationshipViewModel;
 import org.dhis2.commons.data.tuples.Pair;
 import org.dhis2.commons.data.tuples.Trio;
 import org.dhis2.commons.date.DateUtils;
@@ -36,6 +33,9 @@ import org.dhis2.form.ui.validation.FieldErrorMessageProvider;
 import org.dhis2.metadata.usecases.FileResourceConfiguration;
 import org.dhis2.metadata.usecases.ProgramConfiguration;
 import org.dhis2.metadata.usecases.TrackedEntityInstanceConfiguration;
+import org.dhis2.tracker.relationships.model.RelationshipDirection;
+import org.dhis2.tracker.relationships.model.RelationshipOwnerType;
+import org.dhis2.tracker.relationships.model.RelationshipViewModel;
 import org.dhis2.ui.ThemeManager;
 import org.dhis2.usescases.teiDownload.TeiDownloader;
 import org.dhis2.utils.ValueUtils;
