+++ app/src/main/java/org/dhis2/usescases/teiDashboard/dashboardfragments/relationships/RelationshipMapsRepository.kt
@@ -1,7 +1,7 @@
 package org.dhis2.usescases.teiDashboard.dashboardfragments.relationships
 
-import org.dhis2.commons.data.RelationshipOwnerType
 import org.dhis2.maps.model.RelatedInfo
+import org.dhis2.tracker.relationships.model.RelationshipOwnerType
 
 interface RelationshipMapsRepository {
     fun getEventProgram(eventUid: String?): String
