+++ app/src/main/java/org/dhis2/usescases/teiDashboard/dashboardfragments/relationships/RelationshipMapsRepositoryImpl.kt
@@ -2,8 +2,8 @@ package org.dhis2.usescases.teiDashboard.dashboardfragments.relationships
 
 import org.dhis2.commons.bindings.event
 import org.dhis2.commons.bindings.program
-import org.dhis2.commons.data.RelationshipOwnerType
 import org.dhis2.maps.model.RelatedInfo
+import org.dhis2.tracker.relationships.model.RelationshipOwnerType
 import org.dhis2.usescases.events.EventInfoProvider
 import org.dhis2.usescases.tracker.TrackedEntityInstanceInfoProvider
 import org.hisp.dhis.android.core.D2
