+++ app/src/test/java/org/dhis2/uicomponents/map/mocks/RelationshipUiCompomentDummy.kt
@@ -1,6 +1,5 @@
 package org.dhis2.uicomponents.map.mocks
 
-import org.dhis2.commons.data.RelationshipOwnerType
 import org.dhis2.maps.model.RelationshipDirection
 import org.dhis2.maps.model.RelationshipUiComponentModel
 import org.dhis2.maps.model.TeiMap
@@ -31,7 +30,6 @@ object RelationshipUiCompomentDummy {
                 -1,
                 "",
             ),
-            RelationshipOwnerType.TEI,
         )
     }
 
@@ -59,7 +57,6 @@ object RelationshipUiCompomentDummy {
                 -1,
                 "",
             ),
-            RelationshipOwnerType.TEI,
         )
     }
 
@@ -75,7 +72,6 @@ object RelationshipUiCompomentDummy {
             false,
             TeiMap(TEIUID_FROM, geometryFrom, "", -1, ""),
             TeiMap(TEIUID_TO, geometryTo, "", -1, ""),
-            RelationshipOwnerType.TEI,
         )
     }
