+++ dhis2_android_maps/src/main/java/org/dhis2/maps/model/RelationshipUiComponentModel.kt
@@ -1,7 +1,6 @@
 package org.dhis2.maps.model
 
 import org.dhis2.commons.data.CarouselItemModel
-import org.dhis2.commons.data.RelationshipOwnerType
 import org.hisp.dhis.android.core.common.FeatureType
 import org.hisp.dhis.android.core.common.Geometry
 
@@ -13,7 +12,6 @@ data class RelationshipUiComponentModel(
     val bidirectional: Boolean?,
     val from: TeiMap,
     val to: TeiMap,
-    val relationshipOwner: RelationshipOwnerType,
     val canBeDeleted: Boolean? = true,
 ) : CarouselItemModel {
     override fun uid(): String = relationshipUid
