+++ tracker/src/main/kotlin/org/dhis2/tracker/relationships/domain/GetRelationshipsByType.kt
@@ -2,11 +2,11 @@ package org.dhis2.tracker.relationships.domain
 
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.combine
-import org.dhis2.commons.data.RelationshipViewModel
 import org.dhis2.commons.date.DateLabelProvider
 import org.dhis2.tracker.relationships.data.RelationshipsRepository
 import org.dhis2.tracker.relationships.model.RelationshipItem
 import org.dhis2.tracker.relationships.model.RelationshipSection
+import org.dhis2.tracker.relationships.model.RelationshipViewModel
 import org.dhis2.tracker.ui.AvatarProvider
 
 /*
