+++ tracker/src/main/kotlin/org/dhis2/tracker/relationships/model/RelationshipItem.kt
@@ -1,6 +1,5 @@
 package org.dhis2.tracker.relationships.model
 
-import org.dhis2.commons.data.RelationshipOwnerType
 import org.dhis2.ui.avatar.AvatarProviderConfiguration
 
 data class RelationshipItem(
