+++ tracker/src/main/kotlin/org/dhis2/tracker/relationships/ui/RelationshipsScreen.kt
@@ -35,9 +35,9 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import org.dhis2.commons.data.RelationshipOwnerType
 import org.dhis2.tracker.R
 import org.dhis2.tracker.relationships.model.RelationshipItem
+import org.dhis2.tracker.relationships.model.RelationshipOwnerType
 import org.dhis2.tracker.relationships.model.RelationshipSection
 import org.dhis2.ui.avatar.AvatarProvider
 import org.dhis2.ui.avatar.AvatarProviderConfiguration
